优化代码 升级simbot版本至RC.2 升级simbot-mirai组件版本至M4 修改部分代码结构 新增SpringContextUtil获取容器中的bean 修复部分错误

+++ src/main/kotlin/com/huahua/robot/core/common/ListenerAspect.kt
@@ -59,7 +59,7 @@ class ListenerAspect {
                 // 判断是否有权限
                 if (
                     annotation.permission != RobotPermission.MEMBER &&
-                    annotation.permission > role && !RobotCore.isBotAdministrator(author.id.toString())
+                    annotation.permission > role && !RobotCore.isBotAdministrator(author.id.toString().toInt())
                 ) {
                     if (annotation.noPermissionTip.isNotBlank()) {
                         Sender.sendGroupMsg(group, annotation.noPermissionTip)
